copy/dl for grid works, dl for image works
